oee drill down in financial metrics value difference issue solved. and code structure is changed adding new controller for formula calculation
